libwayland-shim: don't require glib
